refactor: add __init__, fix imports in other files
